test: put live Sepolia suite behind //go:build integration (#727)
