+++ core/taskengine/tenderly_client_live_integration_test.go
@@ -1,0 +1,648 @@
+//go:build integration
+// +build integration
+
+package taskengine
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"math/big"
+	"testing"
+	"time"
+
+	"github.com/AvaProtocol/EigenLayer-AVS/core/testutil"
+	avsproto "github.com/AvaProtocol/EigenLayer-AVS/protobuf"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestTenderlyEventSimulation_EndToEnd_Integration(t *testing.T) {
+
+	logger := testutil.GetLogger()
+
+	// Validate Tenderly credentials are configured
+	// These will panic with clear error messages if not configured
+	_ = testutil.GetTestTenderlyAccount()
+	_ = testutil.GetTestTenderlyProject()
+	_ = testutil.GetTestTenderlyAccessKey()
+
+	// Create TenderlyClient with test config
+	config := testutil.GetAggregatorConfig()
+	tenderlyClient := NewTenderlyClient(config, logger)
+
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+
+	t.Run("Basic AnswerUpdated Simulation", func(t *testing.T) {
+		// Create query for Chainlink ETH/USD feed
+		query := &avsproto.EventTrigger_Query{
+			Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+			Topics:    []string{ANSWER_UPDATED_SIG},
+		}
+
+		fmt.Printf("\n=== TENDERLY END-TO-END TEST: Basic AnswerUpdated Simulation ===\n")
+		fmt.Printf("Contract: %s (Sepolia ETH/USD)\n", TENDERLY_SEPOLIA_ETH_USD_FEED)
+		fmt.Printf("Event: AnswerUpdated\n")
+		fmt.Printf("Mode: Real Tenderly API call\n\n")
+
+		// Execute simulation
+		simulatedLog, err := tenderlyClient.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+
+		require.NoError(t, err, "Tenderly simulation should succeed")
+		require.NotNil(t, simulatedLog, "Simulated log should not be nil")
+
+		// Validate the simulated log structure
+		assert.Equal(t, TENDERLY_SEPOLIA_ETH_USD_FEED, simulatedLog.Address.Hex(), "Contract address should match")
+		assert.Len(t, simulatedLog.Topics, 3, "AnswerUpdated should have 3 topics")
+		assert.Equal(t, ANSWER_UPDATED_SIG, simulatedLog.Topics[0].Hex(), "First topic should be AnswerUpdated signature")
+
+		fmt.Printf("Simulation successful.\n")
+		printSimulatedLog(simulatedLog)
+	})
+
+	t.Run("Conditional Price Alert Simulation", func(t *testing.T) {
+		// Test with price > $2000 condition
+		query := &avsproto.EventTrigger_Query{
+			Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+			Topics:    []string{ANSWER_UPDATED_SIG},
+			Conditions: []*avsproto.EventCondition{
+				{
+					FieldName: "current",
+					Operator:  "gt",
+					Value:     "200000000000", // $2000 with 8 decimals
+					FieldType: "int256",
+				},
+			},
+		}
+
+		fmt.Printf("\n=== CONDITIONAL SIMULATION: Price > $2000 ===\n")
+
+		simulatedLog, err := tenderlyClient.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+
+		require.NoError(t, err, "Conditional simulation should succeed")
+		require.NotNil(t, simulatedLog, "Simulated log should not be nil")
+
+		// Validate that the simulated price satisfies the condition
+		priceHex := simulatedLog.Topics[1].Hex()
+		fmt.Printf("Simulated price (hex): %s\n", priceHex)
+
+		// The price should be > $2000 (200000000000 in 8-decimal format)
+		assert.Equal(t, TENDERLY_SEPOLIA_ETH_USD_FEED, simulatedLog.Address.Hex())
+
+		fmt.Printf("Conditional simulation successful.\n")
+		printSimulatedLog(simulatedLog)
+	})
+
+	t.Run("Real Integration with Engine", func(t *testing.T) {
+		// Test the full integration through the Engine
+		fmt.Printf("\n=== FULL ENGINE INTEGRATION TEST ===\n")
+
+		// Create test engine
+		db := testutil.TestMustDB()
+		config := testutil.GetAggregatorConfig()
+		engine := New(db, config, nil, logger)
+
+		// Simulate runTrigger call with Tenderly - use a condition that should match
+		// Set threshold very low to ensure the condition is satisfied
+		triggerConfig := map[string]interface{}{
+			"simulationMode": true,
+			"queries": []interface{}{
+				map[string]interface{}{
+					"addresses": []interface{}{TENDERLY_SEPOLIA_ETH_USD_FEED},
+					"topics": []interface{}{
+						ANSWER_UPDATED_SIG,
+					},
+					"contractAbi": []interface{}{
+						map[string]interface{}{
+							"anonymous": false,
+							"inputs": []interface{}{
+								map[string]interface{}{"indexed": true, "internalType": "int256", "name": "current", "type": "int256"},
+								map[string]interface{}{"indexed": true, "internalType": "uint256", "name": "roundId", "type": "uint256"},
+								map[string]interface{}{"indexed": false, "internalType": "uint256", "name": "updatedAt", "type": "uint256"},
+							},
+							"name": "AnswerUpdated",
+							"type": "event",
+						},
+					},
+					"conditions": []interface{}{
+						map[string]interface{}{
+							"fieldName": "AnswerUpdated.current", // Use nested field path for new data structure
+							"operator":  "gt",
+							"value":     "100000000", // $1.00 - very low threshold to ensure match
+							"fieldType": "int256",
+						},
+					},
+				},
+			},
+		}
+
+		result, err := engine.runEventTriggerImmediately(triggerConfig, map[string]interface{}{})
+
+		require.NoError(t, err, "Engine simulation should succeed")
+
+		// The result can be nil if conditions are not met or simulation fails
+		// This is expected behavior for Tenderly simulation
+		if result == nil {
+			fmt.Printf("No event simulated (conditions not met or simulation failed)\n")
+			fmt.Printf("This can happen when:\n")
+			fmt.Printf("   - Current price doesn't meet the condition\n")
+			fmt.Printf("   - Tenderly API is unavailable\n")
+			fmt.Printf("   - Network connectivity issues\n")
+
+			// Test passes - nil result is valid for failed simulation
+			return
+		}
+
+		// If we get a result, validate its structure
+		assert.True(t, result["success"].(bool), "Should find simulated event")
+
+		// Check if we have the new consistent format: parsed fields in data, raw log in metadata
+		if eventData, hasData := result["data"].(map[string]interface{}); hasData && eventData != nil {
+			// New nested format: parsed ABI fields nested under event name
+			if answerUpdatedData, hasAnswerUpdated := eventData["AnswerUpdated"].(map[string]interface{}); hasAnswerUpdated {
+				assert.NotNil(t, answerUpdatedData["current"], "Should have parsed current price")
+				assert.NotNil(t, answerUpdatedData["roundId"], "Should have parsed round ID")
+				assert.NotNil(t, answerUpdatedData["updatedAt"], "Should have parsed updated timestamp")
+				fmt.Printf("✅ New nested format detected: event fields under 'AnswerUpdated' key\n")
+			} else {
+				t.Errorf("Expected AnswerUpdated event data in nested format")
+			}
+
+			// Check metadata contains raw blockchain log fields
+			if metadata, hasMetadata := result["metadata"].(map[string]interface{}); hasMetadata && metadata != nil {
+				assert.NotNil(t, metadata["address"], "Should have contract address in metadata")
+				assert.NotNil(t, metadata["blockNumber"], "Should have block number in metadata")
+				assert.NotNil(t, metadata["topics"], "Should have topics in metadata")
+				assert.NotNil(t, metadata["data"], "Should have raw data in metadata")
+				fmt.Printf("✅ New nested format detected: parsed fields nested under event name, raw log in metadata\n")
+			} else {
+				t.Errorf("Should have metadata with raw blockchain log data")
+			}
+		} else if evmLog, hasEvmLog := result["evm_log"]; hasEvmLog && evmLog != nil {
+			// Legacy format: evm_log structure
+			assert.NotNil(t, evmLog, "Should have evm_log")
+			fmt.Printf("✅ Legacy evm_log format detected\n")
+		} else {
+			t.Errorf("Result should have either 'data' (new format) or 'evm_log' (legacy format)")
+		}
+
+		// Check common fields
+		assert.NotNil(t, result["success"], "Should have 'success' field")
+
+		fmt.Printf("Full engine integration successful.\n")
+		printEngineResult(result)
+	})
+}
+
+// Helper function to print simulated log details
+func printSimulatedLog(log *types.Log) {
+	fmt.Printf("\n📋 SIMULATED LOG DETAILS:\n")
+	fmt.Printf("Contract: %s\n", log.Address.Hex())
+	fmt.Printf("Block: %d\n", log.BlockNumber)
+	fmt.Printf("TX Hash: %s\n", log.TxHash.Hex())
+	fmt.Printf("TX Index: %d\n", log.TxIndex)
+	fmt.Printf("Log Index: %d\n", log.Index)
+	fmt.Printf("Removed: %t\n", log.Removed)
+
+	fmt.Printf("\nTopics:\n")
+	for i, topic := range log.Topics {
+		fmt.Printf("  [%d] %s", i, topic.Hex())
+		if i == 0 {
+			fmt.Printf(" (AnswerUpdated signature)")
+		} else if i == 1 {
+			fmt.Printf(" (current price)")
+		} else if i == 2 {
+			fmt.Printf(" (round ID)")
+		}
+		fmt.Printf("\n")
+	}
+
+	fmt.Printf("\nData: %s (updatedAt timestamp)\n", "0x"+common.Bytes2Hex(log.Data))
+}
+
+// Helper function to print engine result
+func printEngineResult(result map[string]interface{}) {
+	fmt.Printf("\n📊 ENGINE RESULT:\n")
+	resultJSON, _ := json.MarshalIndent(result, "", "  ")
+	fmt.Printf("%s\n", string(resultJSON))
+
+	if evmLog, ok := result["evm_log"].(map[string]interface{}); ok {
+		fmt.Printf("\n🔍 EVM LOG ANALYSIS:\n")
+		fmt.Printf("Address: %s\n", evmLog["contractAddress"])
+		fmt.Printf("Block: %v\n", evmLog["blockNumber"])
+		fmt.Printf("TX Hash: %s\n", evmLog["transactionHash"])
+
+		if topics, ok := evmLog["topics"].([]string); ok {
+			fmt.Printf("\nTopics Breakdown:\n")
+			for i, topic := range topics {
+				fmt.Printf("  [%d] %s", i, topic)
+				if i == 0 {
+					fmt.Printf(" (AnswerUpdated)")
+				} else if i == 1 {
+					fmt.Printf(" (price)")
+				} else if i == 2 {
+					fmt.Printf(" (roundId)")
+				}
+				fmt.Printf("\n")
+			}
+		}
+	}
+}
+
+// Benchmark the simulation performance
+
+func BenchmarkTenderlySimulation(b *testing.B) {
+	_ = testutil.GetTestTenderlyAccount()
+	_ = testutil.GetTestTenderlyProject()
+	_ = testutil.GetTestTenderlyAccessKey()
+
+	logger := testutil.GetLogger()
+	client := NewTenderlyClient(testutil.GetAggregatorConfig(), logger)
+
+	query := &avsproto.EventTrigger_Query{
+		Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+		Topics:    []string{ANSWER_UPDATED_SIG},
+	}
+
+	ctx := context.Background()
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_, err := client.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+		if err != nil {
+			b.Fatalf("Simulation failed: %v", err)
+		}
+	}
+}
+
+func TestTenderlySimulation_WithConditions_ComprehensiveTest_Integration(t *testing.T) {
+	_ = testutil.GetTestTenderlyAccount()
+	_ = testutil.GetTestTenderlyProject()
+	_ = testutil.GetTestTenderlyAccessKey()
+
+	logger := testutil.GetLogger()
+	client := NewTenderlyClient(testutil.GetAggregatorConfig(), logger)
+
+	ctx := context.Background()
+
+	// First, get the current real price from Tenderly to use in our tests
+	t.Run("GetCurrentPriceData", func(t *testing.T) {
+		_ = testutil.GetTestTenderlyAccount()
+		_ = testutil.GetTestTenderlyProject()
+		_ = testutil.GetTestTenderlyAccessKey()
+		t.Logf("🔗 Using Tenderly HTTP API (RPC gateway deprecated)")
+
+		roundData, err := client.getLatestRoundData(ctx, TENDERLY_SEPOLIA_ETH_USD_FEED, TENDERLY_SEPOLIA_CHAIN_ID)
+		require.NoError(t, err, "Should get real price data from Tenderly")
+		require.NotNil(t, roundData)
+
+		currentPriceFloat := float64(roundData.Answer.Int64()) / 100000000 // Convert to USD
+
+		t.Logf("📊 CURRENT REAL CHAINLINK DATA:")
+		t.Logf("   Contract: %s", TENDERLY_SEPOLIA_ETH_USD_FEED)
+		t.Logf("   Current Price: $%.2f (raw: %s)", currentPriceFloat, roundData.Answer.String())
+		t.Logf("   Round ID: %s", roundData.RoundId.String())
+		t.Logf("   Updated At: %s", time.Unix(roundData.UpdatedAt.Int64(), 0).Format(time.RFC3339))
+
+		// Store current price for use in subsequent tests
+		ctx = context.WithValue(ctx, "currentPrice", roundData.Answer)
+		ctx = context.WithValue(ctx, "currentPriceFloat", currentPriceFloat)
+	})
+
+	// Test 1: Condition that SHOULD match (price > very low threshold)
+	t.Run("ConditionShouldMatch_GreaterThan", func(t *testing.T) {
+		_ = testutil.GetTestTenderlyAccount()
+		_ = testutil.GetTestTenderlyProject()
+		_ = testutil.GetTestTenderlyAccessKey()
+		currentPriceFloat := ctx.Value("currentPriceFloat").(float64)
+
+		// Set threshold much lower than current price to ensure it matches
+		thresholdFloat := currentPriceFloat - 500.0       // $500 below current price
+		thresholdRaw := int64(thresholdFloat * 100000000) // Convert to 8-decimal format
+
+		query := &avsproto.EventTrigger_Query{
+			Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+			Topics:    []string{ANSWER_UPDATED_SIG},
+			Conditions: []*avsproto.EventCondition{
+				{
+					FieldName: "current",
+					Operator:  "gt",
+					Value:     fmt.Sprintf("%d", thresholdRaw),
+					FieldType: "int256",
+				},
+			},
+		}
+
+		t.Logf("TESTING CONDITION THAT SHOULD MATCH:")
+		t.Logf("   Current Price: $%.2f", currentPriceFloat)
+		t.Logf("   Condition: price > $%.2f", thresholdFloat)
+		t.Logf("   Expected: MATCH ✅")
+
+		simulatedLog, err := client.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+		require.NoError(t, err, "Simulation should succeed when condition should match")
+		require.NotNil(t, simulatedLog)
+
+		// Verify the condition was satisfied
+		simulatedPrice := simulatedLog.Topics[1].Big()
+		simulatedPriceFloat := float64(simulatedPrice.Int64()) / 100000000
+
+		assert.True(t, simulatedPrice.Cmp(big.NewInt(thresholdRaw)) > 0,
+			"Simulated price should be greater than threshold")
+
+		t.Logf("CONDITION MATCHED:")
+		t.Logf("   Simulated Price: $%.2f (raw: %s)", simulatedPriceFloat, simulatedPrice.String())
+		t.Logf("   Threshold: $%.2f (raw: %d)", thresholdFloat, thresholdRaw)
+		t.Logf("   Condition Satisfied: %s > %d ✅", simulatedPrice.String(), thresholdRaw)
+
+		// Show complete raw event structure
+		t.Logf("\nRAW EVENT STRUCTURE:")
+		t.Logf("   Address: %s", simulatedLog.Address.Hex())
+		t.Logf("   Block Number: %d", simulatedLog.BlockNumber)
+		t.Logf("   Transaction Hash: %s", simulatedLog.TxHash.Hex())
+		t.Logf("   Topics[0] (Event Sig): %s", simulatedLog.Topics[0].Hex())
+		t.Logf("   Topics[1] (Price): %s", simulatedLog.Topics[1].Hex())
+		t.Logf("   Topics[2] (Round ID): %s", simulatedLog.Topics[2].Hex())
+		t.Logf("   Data (Updated At): 0x%s", common.Bytes2Hex(simulatedLog.Data))
+	})
+
+	// Test 2: Condition that SHOULD NOT match (price > very high threshold)
+	t.Run("ConditionShouldNotMatch_GreaterThan", func(t *testing.T) {
+		_ = testutil.GetTestTenderlyAccount()
+		_ = testutil.GetTestTenderlyProject()
+		_ = testutil.GetTestTenderlyAccessKey()
+		currentPriceFloat := ctx.Value("currentPriceFloat").(float64)
+
+		// Set threshold much higher than current price to test non-matching
+		thresholdFloat := currentPriceFloat + 2000.0      // $2000 above current price
+		thresholdRaw := int64(thresholdFloat * 100000000) // Convert to 8-decimal format
+
+		query := &avsproto.EventTrigger_Query{
+			Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+			Topics:    []string{ANSWER_UPDATED_SIG},
+			Conditions: []*avsproto.EventCondition{
+				{
+					FieldName: "current",
+					Operator:  "gt",
+					Value:     fmt.Sprintf("%d", thresholdRaw),
+					FieldType: "int256",
+				},
+			},
+		}
+
+		t.Logf("TESTING CONDITION THAT SHOULD NOT MATCH:")
+		t.Logf("   Current Price: $%.2f", currentPriceFloat)
+		t.Logf("   Condition: price > $%.2f", thresholdFloat)
+		t.Logf("   Expected: REAL BEHAVIOR - Return real data that doesn't satisfy condition")
+
+		simulatedLog, err := client.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+		require.NoError(t, err, "Simulation should still succeed")
+		require.NotNil(t, simulatedLog)
+
+		// Current behavior: Tenderly returns real current price (not artificial data)
+		simulatedPrice := simulatedLog.Topics[1].Big()
+		simulatedPriceFloat := float64(simulatedPrice.Int64()) / 100000000
+
+		t.Logf("✅ REAL BEHAVIOR - ACTUAL CURRENT PRICE RETURNED:")
+		t.Logf("   Real Current Price: $%.2f", currentPriceFloat)
+		t.Logf("   Simulated Price: $%.2f (raw: %s)", simulatedPriceFloat, simulatedPrice.String())
+		t.Logf("   Threshold: $%.2f (raw: %d)", thresholdFloat, thresholdRaw)
+		t.Logf("   Result: Tenderly returned real price data (not artificial)")
+
+		// The current implementation returns real price data, which should NOT satisfy the high threshold
+		// We expect the real price to be less than the artificially high threshold
+		conditionSatisfied := simulatedPrice.Cmp(big.NewInt(thresholdRaw)) > 0
+
+		if conditionSatisfied {
+			t.Logf("UNEXPECTED: Real price actually satisfies the high threshold!")
+			t.Logf("   This means the current ETH price is > $%.2f", thresholdFloat)
+		} else {
+			t.Logf("EXPECTED: Real price does not satisfy the high threshold")
+			t.Logf("   Real price $%.2f < threshold $%.2f", simulatedPriceFloat, thresholdFloat)
+		}
+
+		// Assert that the simulated price is close to the real current price
+		// Allow for small differences due to timing or data source variations
+		priceDifference := simulatedPriceFloat - currentPriceFloat
+		if priceDifference < 0 {
+			priceDifference = -priceDifference
+		}
+
+		// Price should be within $100 of the real current price (allowing for market movements)
+		assert.True(t, priceDifference < 100.0,
+			"Simulated price should be close to real current price (within $100)")
+
+		t.Logf("\nIMPLEMENTATION NOTE:")
+		t.Logf("   Tenderly simulation returns REAL current price data")
+		t.Logf("   It does NOT generate artificial data to satisfy conditions")
+		t.Logf("   This is more realistic for testing real-world scenarios")
+	})
+
+	// Test 3: Multiple conditions
+	t.Run("MultipleConditions_Complex", func(t *testing.T) {
+		currentPriceFloat := ctx.Value("currentPriceFloat").(float64)
+
+		// Create range: current price ± $100
+		minThreshold := currentPriceFloat - 100.0
+		maxThreshold := currentPriceFloat + 100.0
+		minThresholdRaw := int64(minThreshold * 100000000)
+		maxThresholdRaw := int64(maxThreshold * 100000000)
+
+		query := &avsproto.EventTrigger_Query{
+			Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+			Topics:    []string{ANSWER_UPDATED_SIG},
+			Conditions: []*avsproto.EventCondition{
+				{
+					FieldName: "current",
+					Operator:  "gt",
+					Value:     fmt.Sprintf("%d", minThresholdRaw),
+					FieldType: "int256",
+				},
+				{
+					FieldName: "current",
+					Operator:  "lt",
+					Value:     fmt.Sprintf("%d", maxThresholdRaw),
+					FieldType: "int256",
+				},
+			},
+		}
+
+		t.Logf("TESTING MULTIPLE CONDITIONS (RANGE):")
+		t.Logf("   Current Price: $%.2f", currentPriceFloat)
+		t.Logf("   Condition 1: price > $%.2f", minThreshold)
+		t.Logf("   Condition 2: price < $%.2f", maxThreshold)
+		t.Logf("   Expected: Price in range [$%.2f, $%.2f]", minThreshold, maxThreshold)
+
+		simulatedLog, err := client.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+		require.NoError(t, err, "Multi-condition simulation should succeed")
+		require.NotNil(t, simulatedLog)
+
+		simulatedPrice := simulatedLog.Topics[1].Big()
+		simulatedPriceFloat := float64(simulatedPrice.Int64()) / 100000000
+
+		// Check if both conditions are satisfied
+		condition1Met := simulatedPrice.Cmp(big.NewInt(minThresholdRaw)) > 0
+		condition2Met := simulatedPrice.Cmp(big.NewInt(maxThresholdRaw)) < 0
+
+		t.Logf("MULTIPLE CONDITIONS RESULT:")
+		t.Logf("   Simulated Price: $%.2f", simulatedPriceFloat)
+		t.Logf("   Condition 1 (> $%.2f): %t", minThreshold, condition1Met)
+		t.Logf("   Condition 2 (< $%.2f): %t", maxThreshold, condition2Met)
+		t.Logf("   Both Satisfied: %t", condition1Met && condition2Met)
+	})
+
+	// Test 4: Raw data structure analysis
+	t.Run("RawDataStructureAnalysis", func(t *testing.T) {
+		// Simple query without conditions to see pure simulated data
+		query := &avsproto.EventTrigger_Query{
+			Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+			Topics:    []string{ANSWER_UPDATED_SIG},
+		}
+
+		t.Logf("ANALYZING RAW SIMULATION DATA STRUCTURE:")
+
+		simulatedLog, err := client.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+		require.NoError(t, err)
+		require.NotNil(t, simulatedLog)
+
+		// Decode all fields for analysis
+		price := simulatedLog.Topics[1].Big()
+		roundId := simulatedLog.Topics[2].Big()
+		updatedAtBytes := simulatedLog.Data
+		updatedAt := new(big.Int).SetBytes(updatedAtBytes[len(updatedAtBytes)-32:])
+
+		t.Logf("\nCOMPLETE EVENT BREAKDOWN:")
+		t.Logf("   === Event Metadata ===")
+		t.Logf("   Contract Address: %s", simulatedLog.Address.Hex())
+		t.Logf("   Block Number: %d", simulatedLog.BlockNumber)
+		t.Logf("   Block Hash: %s", simulatedLog.BlockHash.Hex())
+		t.Logf("   Transaction Hash: %s", simulatedLog.TxHash.Hex())
+		t.Logf("   Transaction Index: %d", simulatedLog.TxIndex)
+		t.Logf("   Log Index: %d", simulatedLog.Index)
+		t.Logf("   Removed: %t", simulatedLog.Removed)
+
+		t.Logf("\n   === AnswerUpdated Event Data ===")
+		t.Logf("   Event Signature: %s", simulatedLog.Topics[0].Hex())
+		t.Logf("   Price (current): %s ($%.2f)", price.String(), float64(price.Int64())/100000000)
+		t.Logf("   Round ID: %s", roundId.String())
+		t.Logf("   Updated At: %s (%s)", updatedAt.String(), time.Unix(updatedAt.Int64(), 0).Format(time.RFC3339))
+
+		t.Logf("\n   === Raw Hex Data ===")
+		t.Logf("   Topics[0]: %s (AnswerUpdated signature)", simulatedLog.Topics[0].Hex())
+		t.Logf("   Topics[1]: %s (price as bytes32)", simulatedLog.Topics[1].Hex())
+		t.Logf("   Topics[2]: %s (roundId as bytes32)", simulatedLog.Topics[2].Hex())
+		t.Logf("   Data: 0x%s (updatedAt timestamp)", common.Bytes2Hex(simulatedLog.Data))
+
+		// Verify the data matches expected AnswerUpdated event structure
+		assert.Equal(t, ANSWER_UPDATED_SIG, simulatedLog.Topics[0].Hex(), "First topic should be AnswerUpdated signature")
+		assert.Len(t, simulatedLog.Topics, 3, "Should have exactly 3 topics")
+		assert.Len(t, simulatedLog.Data, 32, "Data should be 32 bytes for updatedAt")
+	})
+}
+
+// Test the proposed enhanced behavior for non-matching conditions
+
+func TestTenderlySimulation_EnhancedConditionHandling_REAL_Integration(t *testing.T) {
+	_ = testutil.GetTestTenderlyAccount()
+	_ = testutil.GetTestTenderlyProject()
+	_ = testutil.GetTestTenderlyAccessKey()
+
+	logger := testutil.GetLogger()
+	client := NewTenderlyClient(testutil.GetAggregatorConfig(), logger)
+
+	ctx := context.Background()
+
+	t.Run("EnhancedBehavior_WithConditionsThatDontMatch", func(t *testing.T) {
+		// First get real current price
+		roundData, err := client.getLatestRoundData(ctx, TENDERLY_SEPOLIA_ETH_USD_FEED, TENDERLY_SEPOLIA_CHAIN_ID)
+		require.NoError(t, err)
+
+		currentPriceFloat := float64(roundData.Answer.Int64()) / 100000000
+
+		// Set a condition that definitely won't match (price > current + $3000)
+		impossibleThreshold := currentPriceFloat + 3000.0
+		impossibleThresholdRaw := int64(impossibleThreshold * 100000000)
+
+		query := &avsproto.EventTrigger_Query{
+			Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+			Topics:    []string{ANSWER_UPDATED_SIG},
+			Conditions: []*avsproto.EventCondition{
+				{
+					FieldName: "current",
+					Operator:  "gt",
+					Value:     fmt.Sprintf("%d", impossibleThresholdRaw),
+					FieldType: "int256",
+				},
+			},
+		}
+
+		t.Logf("🧪 TESTING ENHANCED CONDITION HANDLING:")
+		t.Logf("   Real Current Price: $%.2f", currentPriceFloat)
+		t.Logf("   Impossible Condition: price > $%.2f", impossibleThreshold)
+		t.Logf("   Expected: Return real data with condition_met=false")
+
+		// Test the current method (enhanced method was removed for simplicity)
+		simulatedLog, err := client.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+		if err != nil {
+			t.Logf("⚠️  Current implementation returns error for impossible conditions: %s", err.Error())
+			t.Skip("Current implementation doesn't support enhanced condition handling")
+		}
+		require.NotNil(t, simulatedLog)
+
+		// Extract price from simulated log
+		eventPrice := simulatedLog.Topics[1].Big()
+		eventPriceFloat := float64(eventPrice.Int64()) / 100000000
+
+		t.Logf("\n✅ ENHANCED SIMULATION RESULT:")
+		t.Logf("   Event Price: $%.2f (raw: %s)", eventPriceFloat, eventPrice.String())
+		t.Logf("   Real Price: $%.2f", currentPriceFloat)
+
+		t.Logf("\n💡 NOTE: Current simplified implementation")
+		t.Logf("   When conditions don't match, returns error (no event)")
+		t.Logf("   When conditions match, returns event with real data")
+		t.Logf("   This matches the new protobuf-compliant design")
+	})
+
+	// Test with condition that DOES match
+	t.Run("EnhancedBehavior_WithConditionsThatMatch", func(t *testing.T) {
+		// Get real current price and set a condition that will match
+		roundData, err := client.getLatestRoundData(ctx, TENDERLY_SEPOLIA_ETH_USD_FEED, TENDERLY_SEPOLIA_CHAIN_ID)
+		require.NoError(t, err)
+
+		currentPriceFloat := float64(roundData.Answer.Int64()) / 100000000
+
+		// Set a condition that will definitely match (price > current - $1000)
+		lowThreshold := currentPriceFloat - 1000.0
+		lowThresholdRaw := int64(lowThreshold * 100000000)
+
+		query := &avsproto.EventTrigger_Query{
+			Addresses: []string{TENDERLY_SEPOLIA_ETH_USD_FEED},
+			Topics:    []string{ANSWER_UPDATED_SIG},
+			Conditions: []*avsproto.EventCondition{
+				{
+					FieldName: "current",
+					Operator:  "gt",
+					Value:     fmt.Sprintf("%d", lowThresholdRaw),
+					FieldType: "int256",
+				},
+			},
+		}
+
+		t.Logf("🎯 TESTING CONDITIONS THAT MATCH:")
+		t.Logf("   Real Current Price: $%.2f", currentPriceFloat)
+		t.Logf("   Easy Condition: price > $%.2f", lowThreshold)
+
+		simulatedLog, err := client.SimulateEventTrigger(ctx, query, TENDERLY_SEPOLIA_CHAIN_ID)
+		require.NoError(t, err)
+
+		// When conditions match, should always use real price
+		eventPrice := simulatedLog.Topics[1].Big()
+		eventPriceFloat := float64(eventPrice.Int64()) / 100000000
+
+		t.Logf("✅ Conditions naturally satisfied - using real price: $%.2f", eventPriceFloat)
+	})
+}
+
+// TestEventConditionEvaluation_Unit tests the actual condition evaluation logic used by the engine
